Add prefetch_size to query options

+++ resource/constants.rs
@@ -32,7 +32,7 @@ pub mod server {
     pub const GRPC_CONNECTION_KEEPALIVE: Duration = Duration::from_secs(2 * SECONDS_IN_HOUR);
 
     // TODO: Maybe we start moving these options to separate crates?
-    pub const DEFAULT_PREFETCH_SIZE: u64 = 32;
+    pub const DEFAULT_PREFETCH_SIZE: usize = 32;
     pub const DEFAULT_SCHEMA_LOCK_ACQUIRE_TIMEOUT_MILLIS: u64 = Duration::from_secs(10).as_millis() as u64;
     pub const DEFAULT_TRANSACTION_TIMEOUT_MILLIS: u64 = Duration::from_secs(5 * SECONDS_IN_MINUTE).as_millis() as u64;
     pub const DEFAULT_TRANSACTION_PARALLEL: bool = true;
